Move calibration curve subclasses inside CalibrationCurve
